user-service-taskly/requirements.txt
Flask==2.0.1
Jinja2==3.0.1
psycopg2-binary==2.9.1
psycopg2==2.9.1 
PyJWT==2.1.0
Pyrebase4==4.5.0